linter passing, fix import cycle
