--- internal/core/models/templates.go
@@ -1,7 +1,0 @@
-package models
-
-type UserDeviceAutoPIUnit struct {
-	UserDeviceID       string
-	DeviceDefinitionID string
-	DeviceStyleID      string
-}
