+++ internal/core/services/user_device_service.go
@@ -5,7 +5,7 @@ import (
 	"encoding/hex"
 	"fmt"
 
-	"github.com/DIMO-Network/vehicle-signal-decoding/internal/core/models"
+	"github.com/DIMO-Network/vehicle-signal-decoding/internal/core/appmodels"
 
 	pb "github.com/DIMO-Network/devices-api/pkg/grpc"
 
@@ -16,7 +16,7 @@ import (
 
 //go:generate mockgen -source user_device_service.go -destination mocks/user_device_service_mock.go
 type UserDeviceService interface {
-	GetUserDeviceServiceByAutoPIUnitID(ctx context.Context, id string) (*models.UserDeviceAutoPIUnit, error)
+	GetUserDeviceServiceByAutoPIUnitID(ctx context.Context, id string) (*appmodels.UserDeviceAutoPIUnit, error)
 	GetUserDeviceByVIN(ctx context.Context, vin string) (*pb.UserDevice, error)
 	GetUserDeviceByEthAddr(ctx context.Context, ethAddr string) (*pb.UserDevice, error)
 }
@@ -31,7 +31,7 @@ func NewUserDeviceService(settings *config.Settings) UserDeviceService {
 	}
 }
 
-func (a *userDeviceService) GetUserDeviceServiceByAutoPIUnitID(ctx context.Context, id string) (*models.UserDeviceAutoPIUnit, error) {
+func (a *userDeviceService) GetUserDeviceServiceByAutoPIUnitID(ctx context.Context, id string) (*appmodels.UserDeviceAutoPIUnit, error) {
 
 	deviceClient, conn, err := a.getDeviceGrpcClient()
 	if err != nil {
@@ -45,7 +45,7 @@ func (a *userDeviceService) GetUserDeviceServiceByAutoPIUnitID(ctx context.Conte
 		return nil, err
 	}
 
-	return &models.UserDeviceAutoPIUnit{
+	return &appmodels.UserDeviceAutoPIUnit{
 		UserDeviceID:       userDevice.UserDeviceId,
 		DeviceDefinitionID: userDevice.DeviceDefinitionId,
 		DeviceStyleID:      userDevice.DeviceStyleId,
